# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SlumSafe CV API
  version: 1.0.0
  description: REST edge API for inspection upload, ranking, dashboard analytics, and risk explainability.
servers:
  - url: https://api.slumsafe.example.com/v1
security:
  - firebaseAuth: []
paths:
  /auth/demo-session:
    post:
      summary: Create a demo operator session for the dashboard or mobile demo mode
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: [Inspector, Supervisor, Admin]
      responses:
        '200':
          description: Demo session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemoSession'
  /cities:
    get:
      summary: List supported city deployments
      tags: [Analytics]
      responses:
        '200':
          description: City list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CityOption'
  /dashboard/summary:
    get:
      summary: Fetch city summary metrics and before-vs-after queue snapshot
      tags: [Analytics]
      parameters:
        - in: query
          name: cityId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummary'
  /inspections:
    post:
      summary: Create an inspection record and issue an upload session
      tags: [Inspection]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInspectionRequest'
      responses:
        '201':
          description: Inspection created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectionResponse'
  /inspections/{inspectionId}/sync:
    post:
      summary: Mark an offline inspection as synchronized
      tags: [Inspection]
      parameters:
        - in: path
          name: inspectionId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sync acknowledged
  /audit/logs:
    get:
      summary: Retrieve audit logs for operational activity
      tags: [Analytics]
      parameters:
        - in: query
          name: cityId
          required: true
          schema:
            type: string
        - in: query
          name: role
          schema:
            type: string
            enum: [Inspector, Supervisor, Admin]
      responses:
        '200':
          description: Audit log feed
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLogItem'
  /risk-queue:
    get:
      summary: Fetch a ranked inspection queue
      tags: [Risk]
      parameters:
        - in: query
          name: cityId
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: role
          schema:
            type: string
            enum: [Inspector, Supervisor, Admin]
      responses:
        '200':
          description: Ranked structures
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/RiskQueueItem'
  /structures/{structureId}/risk:
    get:
      summary: Retrieve risk score, explanation, and evidence
      tags: [Risk]
      parameters:
        - in: path
          name: structureId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Risk detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskDetailResponse'
  /satellite/features/{structureId}:
    post:
      summary: Trigger satellite feature refresh
      tags: [Satellite]
      parameters:
        - in: path
          name: structureId
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Satellite job started
  /reports/export:
    post:
      summary: Export analytics report
      tags: [Analytics]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cityId:
                  type: string
                format:
                  type: string
                  enum: [pdf, csv]
      responses:
        '200':
          description: Export generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportExportResponse'
components:
  securitySchemes:
    firebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CreateInspectionRequest:
      type: object
      required: [cityId, structureId, inspectorId, imageCount, damageFeatures]
      properties:
        cityId:
          type: string
        structureId:
          type: string
        inspectorId:
          type: string
        imageCount:
          type: integer
        damageFeatures:
          type: object
          properties:
            crackCount:
              type: integer
            spallAreaRatio:
              type: number
            leaningSeverity:
              type: number
        captureQuality:
          type: object
          properties:
            blurVariance:
              type: number
            lightingScore:
              type: number
        capturedAt:
          type: string
          format: date-time
    InspectionResponse:
      type: object
      properties:
        inspectionId:
          type: string
        uploadUrl:
          type: string
        status:
          type: string
          enum: [pending_upload, processing, completed]
    DemoSession:
      type: object
      properties:
        token:
          type: string
        user:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            role:
              type: string
              enum: [Inspector, Supervisor, Admin]
            cityIds:
              type: array
              items:
                type: string
    CityOption:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        region:
          type: string
    DashboardSummary:
      type: object
      properties:
        cityId:
          type: string
        cityName:
          type: string
        totalStructures:
          type: integer
        highRiskStructures:
          type: integer
        averageApiResponseMs:
          type: integer
        coverageRate:
          type: number
        prioritizedQueue:
          type: array
          items:
            $ref: '#/components/schemas/QueueSnapshot'
        randomQueue:
          type: array
          items:
            $ref: '#/components/schemas/QueueSnapshot'
    QueueSnapshot:
      type: object
      properties:
        structureId:
          type: string
        riskProbability:
          type: number
    RiskQueueItem:
      type: object
      properties:
        structureId:
          type: string
        locality:
          type: string
        city:
          type: string
        ward:
          type: string
        lat:
          type: number
        lng:
          type: number
        riskProbability:
          type: number
        confidence:
          type: number
        queueRank:
          type: integer
        nextInspectionDueDays:
          type: integer
        topFactors:
          type: array
          items:
            type: string
        assignedLabel:
          type: string
    AuditLogItem:
      type: object
      properties:
        id:
          type: string
        cityId:
          type: string
        actorName:
          type: string
        actorRole:
          type: string
          enum: [Inspector, Supervisor, Admin]
        action:
          type: string
        targetId:
          type: string
        occurredAt:
          type: string
          format: date-time
        outcome:
          type: string
    RiskDetailResponse:
      type: object
      properties:
        structureId:
          type: string
        cityId:
          type: string
        locality:
          type: string
        ward:
          type: string
        riskProbability:
          type: number
        confidence:
          type: number
        reinspectionIntervalDays:
          type: integer
        trend:
          type: array
          items:
            type: number
        topContributors:
          type: array
          items:
            type: object
            properties:
              feature:
                type: string
              direction:
                type: string
                enum: [increase, decrease]
              contribution:
                type: number
        evidence:
          type: object
          properties:
            latestInspectionId:
              type: string
            satelliteStatus:
              type: string
            rainfallMm7d:
              type: number
            footprintSource:
              type: string
    ReportExportResponse:
      type: object
      properties:
        fileName:
          type: string
        contentType:
          type: string
        content:
          type: string